src/main/java/com/keymetrics/service/DeploymentService.java
package com.keymetrics.service;

import com.keymetrics.entity.Metrics;
import com.keymetrics.repository.MetricsRepository;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
import org.springframework.stereotype.Service;

import java.util.Date;

@Service
@RequiredArgsConstructor
@EnableMongoRepositories(basePackages = "com.keymetrics.repository")
@Slf4j
public class DeploymentService {

    private final MetricsRepository metricsRepository;

    public void update(String name, Integer environment) {
        Metrics metrics = new Metrics(name, environment, new Date());
        Metrics save = metricsRepository.save(metrics);
        System.out.println(">>>> saved " + save.deployedAt.toString());
    }
}
